cloud_functions/requirements.txt
# Cloud Functions Requirements
# Python 3.11+ compatible

# Firebase Admin SDK
firebase-admin==6.5.0

# Google Cloud Libraries
functions-framework==3.*
google-cloud-firestore==2.16.0
google-cloud-secret-manager==2.20.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
geopandas==1.0.1
fiona>=1.10.0
shapely==2.0.3
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dateutil==2.9.0
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
